KAIR_nkd/requirement.txt
opencv-python
scikit-image
pillow
torchvision
hdf5storage
ninja
timm